57 fix compatibility with very large inputs in cpp mode

+++ src/cpp/helper_dfm.cc
@@ -18,10 +18,10 @@
  */
 void copy_vec_with_stride(vector<double> &src,
                           double *dest,
-                          size_t start,
-                          size_t stride)
+                          unsigned long long start,
+                          unsigned long long stride)
 {
-    for (size_t ii = 0; ii < src.size(); ii++)
+    for (unsigned long long ii = 0; ii < (unsigned long long)(src.size()); ii++)
     {
         dest[ii * stride + start] = src[ii];
     }
@@ -35,20 +35,20 @@ void copy_vec_with_stride(vector<double> &src,
     Make element contiguous in memory.
  */
 void make_full_shifted_isf(double *vec,
-                           size_t nx,
-                           size_t ny,
-                           size_t nt)
+                           unsigned long long nx,
+                           unsigned long long ny,
+                           unsigned long long nt)
 {
     // FFTshift along y only
-    size_t c = ny / 2;
-    size_t _nx = nx / 2 + 1;
+    unsigned long long c = ny / 2;
+    unsigned long long _nx = nx / 2 + 1;
     if (ny % 2 == 0)
     {
-        for (size_t t = 0; t < nt; t++)
+        for (unsigned long long t = 0; t < nt; t++)
         {
-            for (size_t x = 0; x < _nx; x++)
+            for (unsigned long long x = 0; x < _nx; x++)
             {
-                for (size_t y = 0; y < c; y++)
+                for (unsigned long long y = 0; y < c; y++)
                 {
                     swap(vec[2 * (t * ny * _nx + y * _nx + x)], vec[2 * (t * ny * _nx + (y + c) * _nx + x)]);
                 }
@@ -58,12 +58,12 @@ void make_full_shifted_isf(double *vec,
     else
     {
         double tmp = 0.0;
-        for (size_t t = 0; t < nt; t++)
+        for (unsigned long long t = 0; t < nt; t++)
         {
-            for (size_t x = 0; x < _nx; x++)
+            for (unsigned long long x = 0; x < _nx; x++)
             {
                 tmp = vec[2 * (t * ny * _nx + x)];
-                for (size_t y = 0; y < c; y++)
+                for (unsigned long long y = 0; y < c; y++)
                 {
                     vec[2 * (t * ny * _nx + y * _nx + x)] = vec[2 * (t * ny * _nx + (y + c + 1) * _nx + x)];
                     vec[2 * (t * ny * _nx + (y + c + 1) * _nx + x)] = vec[2 * (t * ny * _nx + (y + 1) * _nx + x)];
@@ -74,11 +74,11 @@ void make_full_shifted_isf(double *vec,
     }
 
     // Collapse elements to the right
-    for (size_t t = 0; t < nt; t++)
+    for (unsigned long long t = 0; t < nt; t++)
     {
-        for (size_t y = 0; y < ny; y++)
+        for (unsigned long long y = 0; y < ny; y++)
         {
-            for (size_t x = 1; x < _nx; x++)
+            for (unsigned long long x = 1; x < _nx; x++)
             {
                 vec[2 * (t * ny * _nx + y * _nx + nx / 2) - x] = vec[2 * (t * ny * _nx + y * _nx + nx / 2) - 2 * x];
             }
@@ -86,20 +86,20 @@ void make_full_shifted_isf(double *vec,
     }
 
     // Make full (copy missing values)
-    for (size_t t = 0; t < nt; t++)
+    for (unsigned long long t = 0; t < nt; t++)
     {
         if (ny % 2 == 0)
         {
             // mirror first row
-            for (size_t x = 0; x < _nx - 1; x++)
+            for (unsigned long long x = 0; x < _nx - 1; x++)
             {
                 vec[2 * (t * ny * _nx) + x] = vec[2 * (t * ny * _nx + _nx - 1) - x];
             }
             // for other rows, make symmetry around center
-            for (size_t y = 1; y < ny; y++)
+            for (unsigned long long y = 1; y < ny; y++)
             {
                 // make symmetry around center
-                for (size_t x = 0; x < _nx - 1; x++)
+                for (unsigned long long x = 0; x < _nx - 1; x++)
                 {
                     vec[2 * (t * ny * _nx + y * _nx) + x] = vec[2 * (t * ny * _nx + (ny - y) * _nx) + 2 * (nx / 2) - x];
                 }
@@ -107,10 +107,10 @@ void make_full_shifted_isf(double *vec,
         }
         else
         {
-            for (size_t y = 0; y < ny; y++)
+            for (unsigned long long y = 0; y < ny; y++)
             {
                 // make symmetry around center
-                for (size_t x = 0; x < _nx - 1; x++)
+                for (unsigned long long x = 0; x < _nx - 1; x++)
                 {
                     vec[2 * (t * ny * _nx + y * _nx) + x] = vec[2 * (t * ny * _nx + (ny - y - 1) * _nx) + 2 * (nx / 2) - x];
                 }
@@ -119,11 +119,11 @@ void make_full_shifted_isf(double *vec,
     }
 
     // Move to front
-    for (size_t t = 0; t < nt; t++)
+    for (unsigned long long t = 0; t < nt; t++)
     {
-        for (size_t y = 0; y < ny; y++)
+        for (unsigned long long y = 0; y < ny; y++)
         {
-            for (size_t x = 0; x < nx; x++)
+            for (unsigned long long x = 0; x < nx; x++)
             {
                 vec[t * ny * nx + y * nx + x] = vec[2 * (t * ny * _nx + y * _nx) + x];
             }
